+++ gems/aws-sdk-core/lib/aws-sdk-core/plugins/retry_errors.rb
@@ -118,6 +118,7 @@ A delay randomiser function used by the default backoff function. Some predefine
 
         def networking?
           @error.is_a?(Seahorse::Client::NetworkingError) ||
+          @error.is_a?(Errors::NoSuchEndpointError) ||
           NETWORKING_ERRORS.include?(@name)
         end